requirements.txt
aniso8601==8.0.0
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-JWT-Extended==3.24.1
flask-marshmallow==0.11.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
#Flask-Uploads==0.2.1 # Need to install this one from install git+git://github.com/maxcountryman/flask-uploads.git@master
idna==2.9
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
marshmallow==3.5.1
marshmallow-sqlalchemy==0.22.3
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pyparsing==2.4.6
pytest==5.4.1
python-dotenv==0.12.0
pytz==2019.3
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.15
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==1.0.0
zipp==3.1.0
